# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface VerifyLoginBody {
  /** Array of letter numbers the user clicked in order */
  sequence: number[];
}

export interface LoginResult {
  success: boolean;
  partial?: boolean;
  message: string;
  /** master | lead | deputy_lead — only present on success */
  role?: string;
  /** Campus name — only present when role is lead or deputy_lead */
  campus?: string;
}

export interface LoginCodeConfig {
  /** The ordered sequence of letter numbers that grant access */
  code: number[];
  updatedAt: string;
}

export interface UpdateLoginCodeBody {
  /** New sequence of letter numbers */
  code: number[];
  /** Admin password to authorize the change */
  adminPassword: string;
}

export interface Worker {
  id: number;
  name: string;
  role?: string;
  category: string;
  campus: string;
  photoUrl?: string;
  onHold: boolean;
  createdAt: string;
}

export interface CreateWorkerBody {
  name: string;
  role?: string;
  category: string;
  campus: string;
  photoUrl?: string;
}

export interface UpdateWorkerBody {
  name?: string;
  role?: string;
  campus?: string;
  photoUrl?: string;
  onHold?: boolean;
}

export interface WorkerList {
  workers: Worker[];
}

export interface CheckIn {
  id: number;
  workerId: number;
  campus: string;
  service: string;
  serviceDate: string;
  checkedInAt: string;
}

export interface CreateCheckInBody {
  workerId: number;
  campus: string;
  service: string;
  serviceDate: string;
}

export interface CheckInList {
  checkIns: CheckIn[];
}

export interface ServiceReport {
  id: number;
  campus: string;
  service: string;
  serviceDate: string;
  totalEntries: number;
  servants: number;
  salvations: number;
  prayers: number;
  family: number;
  notes?: string;
  updatedAt: string;
}

export interface UpsertServiceReportBody {
  campus: string;
  service: string;
  serviceDate: string;
  totalEntries?: number;
  servants?: number;
  salvations?: number;
  prayers?: number;
  family?: number;
  notes?: string;
}

export interface ServiceReportList {
  reports: ServiceReport[];
}

export interface AltarReport {
  id: number;
  name: string;
  campus: string;
  service: string;
  responseType: string;
  phone?: string;
  email?: string;
  notes?: string;
  createdAt: string;
}

export interface CreateAltarReportBody {
  name: string;
  campus: string;
  service: string;
  responseType: string;
  phone?: string;
  email?: string;
  notes?: string;
}

export interface AltarReportList {
  reports: AltarReport[];
}

export interface ServiceNotes {
  date: string;
  service: string;
  notes: string;
}

export interface SaveServiceNotesBody {
  date: string;
  service: string;
  notes: string;
}

export interface DailyAltarReport {
  id: number;
  date: string;
  campus: string;
  service: string;
  salvations: number;
  prayers: number;
  altarMembers: number;
  notes: string;
  createdAt: string;
  updatedAt: string;
}

export interface UpsertDailyAltarReportBody {
  date: string;
  campus: string;
  service: string;
  salvations?: number;
  prayers?: number;
  altarMembers?: number;
  notes?: string;
}

export interface DailyAltarReportList {
  reports: DailyAltarReport[];
}

export interface CampusPasswordEntry {
  campus: string;
  role: string;
  hasPassword: boolean;
  sequence?: number[];
}

export interface CampusPasswordList {
  passwords: CampusPasswordEntry[];
}

export interface SetCampusPasswordBody {
  campus: string;
  role: string;
  /** Hebrew letter number sequence */
  sequence: number[];
}

export interface VerifyCampusPasswordBody {
  campus: string;
  role: string;
  password: string;
}

export interface PasswordHistoryEntry {
  id: number;
  campus: string;
  role: string;
  sequence: number[];
  changedAt: string;
}

export interface PasswordHistoryList {
  entries: PasswordHistoryEntry[];
}

export interface TeamPreset {
  campus: string;
  service: string;
  workerIds: number[];
}

export interface SetTeamPresetBody {
  campus: string;
  service: string;
  workerIds: number[];
}

export interface VerifyCampusPasswordResponse {
  valid: boolean;
}

export interface ActivityLog {
  id: number;
  tool: string;
  action: string;
  accessedAt: string;
}

export interface ActivityLogList {
  logs: ActivityLog[];
}

export interface CreateActivityLogBody {
  tool: string;
  action: string;
}

export type DbancContactCustomData = { [key: string]: unknown };

export interface DbancContact {
  id: number;
  firstName: string;
  lastName: string;
  phone: string;
  carrier: string;
  gender: string;
  campus: string;
  notes: string;
  customData: DbancContactCustomData;
  createdAt: string;
  updatedAt: string;
}

export interface DbancContactList {
  contacts: DbancContact[];
}

export type CreateDbancContactBodyCustomData = { [key: string]: unknown };

export interface CreateDbancContactBody {
  firstName: string;
  lastName: string;
  phone: string;
  carrier?: string;
  gender?: string;
  campus?: string;
  notes?: string;
  customData?: CreateDbancContactBodyCustomData;
}

export interface DbancCustomField {
  id: number;
  label: string;
  fieldType: string;
  options: string[];
  sortOrder: number;
  createdAt: string;
}

export interface DbancCustomFieldList {
  fields: DbancCustomField[];
}

export interface CreateDbancCustomFieldBody {
  label: string;
  fieldType?: string;
  options?: string[];
  sortOrder?: number;
}

export type PxpConfigScriptTree = { [key: string]: unknown };

export interface PxpConfig {
  id: number;
  churchName: string;
  scriptTree: PxpConfigScriptTree;
  updatedAt: string;
}

export type UpdatePxpConfigBodyScriptTree = { [key: string]: unknown };

export interface UpdatePxpConfigBody {
  churchName?: string;
  scriptTree?: UpdatePxpConfigBodyScriptTree;
}

export interface PxpCallLog {
  id: number;
  contactId: number;
  callerName: string;
  campus: string;
  outcome: string;
  notes: string;
  calledAt: string;
}

export interface PxpCallLogList {
  logs: PxpCallLog[];
}

export interface CreatePxpCallLogBody {
  contactId: number;
  callerName: string;
  campus: string;
  outcome?: string;
  notes?: string;
}

export type ListWorkersParams = {
  category?: string;
  campus?: string;
};

export type ListCheckInsParams = {
  campus: string;
  service: string;
  serviceDate: string;
};

export type ListPxpCallLogsParams = {
  contactId?: number;
};

export type ListActivityLogsParams = {
  tool?: string;
};

export type GetTeamPresetParams = {
  campus: string;
  service: string;
};

export type ListServiceReportsParams = {
  campus?: string;
};

export type GetServiceNotesParams = {
  date: string;
  service: string;
};

export type ListDailyAltarReportsParams = {
  month?: number;
  year?: number;
};